Implemented search for decan/term changes in progressing charts..
ChartTransitSearch now recognized -v3 settings for decans and terms
and lists the dates of those changes along with the aspects.
For the feature to take effect the sign changes need to be
unrestricted. Normal sign changes aren't listed with v3 because
they are listed as first part of the decan/term change designator.