Modified CheckSignChange for a common concept what is passed..
as positional value in cases aSign, aDeg, aRul, aDec.
Formerly the value passed for aSign was not consistent with that
for aDeg. Now we have more cases and all share a common concept
what is passed for these transit events, i.e. the border position.
Minor changes in PrintAspect to make it work.