Fixed some problems with the postscript output..
by first changing the code in PsBegin() to postscript 3, taking
the gutter out of the bounding box and adding an %%Orientation
field to it. Then changed the code in DrawClearScreen() so that it
produces landscape (if needed) instead of seascape. Then applied
a reset to all relevant settings before doing the postscript
setup. With this charts of all dimensions fit nicely on postscript
pages in Letter, A4, A5 and so on. If one wants to have special
scalings for the text, there is always the option to print EPS
files which preserve the actual screen settings.